requirements.txt
tensorflow>=2.0.0
opencv-python>=4.5.0
keras-facenet>=0.3.1
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.0.0 